Refactor alarm scheduling code. Create ScheduleExactAlarmPermissionReceiver.kt.
